main/requirements.txt
litellm
python-dotenv
streamlit
spacy
sentence-transformers
scikit-learn